combine Fetch passthrough tests into Fetch.test.ts
Amp-Thread-ID: https://ampcode.com/threads/T-019cb10b-2f8f-7459-9590-0ca2b186da73
